creando repprostori y services de venta

+++ demo/src/main/java/com/SistemaVenta/demo/Model/DetailsSale.java
@@ -31,7 +31,7 @@ public class DetailsSale {
     
     private double precio;
 
-    private double cantidad;
+    private Integer cantidad;
 
     private double subtotal;
 }
